Add serial run script.
This serial script takes the generated potential file, replaces the lattice
parameter, then feeds it into the next calculation, doing everything in serial.

The script takes a very different format of input than the other input script,
which is something that I'll address in new versions of both of these.